Add conditional skipping of IPMI selftests
Some Machines have flaky or buggy IPMI implementations. It can happen,
that the selftest requests to the BMC fails. In some cases we want to
run the tests anyway, for example if we only have one machine of that
architecture and cannot easily replace this.
Simply set IPMI_SKIP_SELFTEST to 1 in order to skip the selftest. By
default, the selftest will be performed.

Signed-off-by: Michael Moese <frankenmichl@posteo.de>