Makefiles: set CXXFLAGS not CFLAGS or CPPFLAGS
CFLAGS is for C compilation. CPPFLAGS is for the preprocessor.
CXXFLAGS is for C++ compilation, which is what we're actually
trying to apply the flags to in all three cases. CPPFLAGS happen
to be applied to the compile stage as well as the preprocessor
usually, so CPPFLAGS was working, but it wasn't correct. CFLAGS
is not applied to C++ compiles, so `videoencoder_CFLAGS` was
not being used at all.

This broke build on Fedora Rawhide, because /usr/include/opencv2
no longer exists there. There is /usr/include/opencv4/opencv2 and
this should be fine because we should get -I/usr/include/opencv4
from OPENCV_CFLAGS, but because of this bug, that wasn't set
correctly and the build failed. With this change it works.

Thanks to Orion Poplowski for reminding me of the CPPFLAGS vs.
CXXFLAGS distinction here.

Signed-off-by: Adam Williamson <awilliam@redhat.com>