CI: Install possibly new dependencies
This way we can add a new dependency and use it in the same PR.

In the future `tools/tools.sh` can also be used within the Dockerfile,
so we don't have to generate the dependencies there.

For that OBS needs to be configured to extract dependencies.yaml and
tools/tools.sh first.

Issue: https://progress.opensuse.org/issues/56525