Fix missing qemu-kvm dependency on non-x86_64
qemu-kvm was added to the spec file in 8f30ade5 but was mentioned in
before already as part of the test Dockerfile. As it was not provided as
requirement in the spec file as runtime dependencies for the generated
packages we can just try to remove it from dependencies and rely on CI
and package build tests.

Related progress issue: https://progress.opensuse.org/issues/67837