Move git to %main_requires as git-core
See:
https://github.com/os-autoinst/os-autoinst/pull/1422#issuecomment-640941787
since that change, it seems the tests fail if git is not present.
That means we need git at both runtime and test time, so we can
just put it in `%main_requires` and remove the concept of
`%requires_not_needed_in_tests` entirely as git was the last
thing in it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>