spec: use 'check-doc' Makefile target, not 'check'
The 'check' Makefile target basically runs code quality checks
then runs the 'check-doc' target (which runs spellchecks) as
long as CHECK_DOC isn't 0. In the RPM spec, we're calling 'make
check', but before that, we are hackily disabling all the code
quality checks, so all the 'check' target winds up doing is
calling 'check-doc' (except for SLE, where we set CHECK_DOC=0
because SLE doesn't have the necessary deps). I think it'd make
more sense to drop use of the 'check' target entirely, and just
call 'check-doc' directly when we want to do the spellchecks.

Signed-off-by: Adam Williamson <awilliam@redhat.com>