Separate lint requires from test requires, don't include in spec
The Critic requirements are not really test requirements. You
don't need them to run `make test`. They are only needed to run
the linting checks in `make check`. We don't run those tests in
the RPM spec, so including these deps in the test requirements
and hence in the RPM spec's BuildRequires is technically wrong.

This splits those requirements into a new dependencies.yaml group
and adjusts the logic a bit to handle this new group.

One wrinkle here is that the 00-compile-check-all.t check finds
the custom Critic policy module we ship under tools/ and tries
to compile check it; if the Critic deps are not available, that
will fail. This deals with that by just excepting that file from
the test. It's not really a part of the installed tool, after
all, so arguably `make test` shouldn't syntax check it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>