Fix test for undocumented backend variables
https://progress.opensuse.org/issues/34303

Problem: When executing the test manually it shows some errors
but the exit code is 0

Solution: Force an exit with 1 error code.

Problem: Some vars don't have description. With this PR this generates the failure
of the test.

Solution: A new description has been added

Problem: When executing the test manually the method write_doc
shows some errors but exit code is 0. And some vars don't have
description or aren't on the file backend_vars.

Solution: Force an exit with 1 error code. And fix or add the
required variables and description.