Add support for TPMv2 emulation via swtpm
This adds support in os-autoinst for running qemu with a TPM
emulator device, using swtpm running on the worker host. See
https://www.qemu.org/docs/master/specs/tpm.html for general
background on this.

All this code does is run qemu with appropriate arguments if
the QEMUTPM setting is set. We use `/tmp/mytpmX/swtpm-sock` as
the socket path, where X is replaced by the worker's instance
number if QEMUTPM is set to 'instance', or by the QEMUTPM value
otherwise.

We explicitly do not handle running and shutting down swtpm
itself; I considered that design but thought it would get too
complex. Instead the admin is expected to handle that themselves,
by using a systemd unit or something similar.

I've tested this on x86_64 and it works. I haven't tested on
the other two arches yet as Fedora is moving data centres so
we're on reduced capacity, and I don't have aarch64 or ppc64le
worker hosts up yet.

Signed-off-by: Adam Williamson <awilliam@redhat.com>