Prevent calling cv::imwrite with an empty image
This change is supposed to prevent

```
terminate called after throwing an instance of 'cv::Exception'
  what():  OpenCV(3.4.10) …/opencv-3.4.10/modules/imgcodecs/src/loadsave.cpp:759: error: (-215:Assertion failed) !_img.empty() in function 'imwrite'
```

The occasion where cv::imwrite is called in videoencoder.cpp can be
safely wrapped around a check whether the image is actually present
without hiding errors.

Wrapping the occasion in the XS interface is likely ok, too. The
function already has a return value which is supposed to be checked
for errors if relevant so an empty image would just be another error.
I also added logging in consistency with image_read.

See https://progress.opensuse.org/issues/68554