Avoid updating last_good if there is no possible user of it
Taking a snapshot is really expensive, so it should be avoided.
If there are two milestone tests and the latter is fatal, a failure causes
the entire run to fail instead of rolling back, so last_good won't be used.
The last test does not need to update last_good either.