os-autoinst-openvswitch: Fix spurious network startup race-conditions (2nd)
Similar as in 05a44404 we need to wait further until also an IP is ready
on the device as the original problem is still present. It looks like
just waiting for the bridge device is not enough to ensure an IP is
configured so we should wait for that explicitly as well.

Tested manually on openqaworker12 with

```
$ ./os-autoinst-openvswitch
Waiting for bridge 'br0' to be created and configured...
Waiting for bridge 'br0' to be created and configured...
…
```

and

```
env OS_AUTOINST_USE_BRIDGE=br1 ./os-autoinst-openvswitch
```

Related progress issue: https://progress.opensuse.org/issues/59300