Suppress 'unable to inform websocket…' error when command server exited
When the command server has already exited we would not do that http
request anyways. This change improves the case  when the command server
exits while the http request is sent (and therefore the http request
fails). Then it makes more sense to suppress the error because the
command server has likely just been stopped by the worker or the user.
A message like `commands process exited: 0` is still logged in any case.