spec file: Do not abuse %{_libexecdir} for '/usr/lib'
In the not too far future %{_libexecdir} will change from '/usr/lib' to
'/usr/libexec'. However, we want to keep installing to
'/usr/lib/os-autoinst' in consistency with Perl itself (which uses
`/usr/lib/perl5`).

The CMake build script already does the right thing but the use of
%{_libexecdir} within the spec file needs to be changed to %{_prefix}/lib.
(Not using %{_libdir} because that would be '/usr/lib64'.)