docker: Add Dockerfile for isotovideo+qemu-x86
I had
https://build.opensuse.org/package/view_file/home:okurz:container/isotovideo/Dockerfile
since about a year. This commit pushes that upstream to os-autoinst and
adapts to work in more general CI pipelines in a limited way, e.g. also
github actions where no nested virt and hence no KVM is available. With
this commit we can publish the container and image as an easier entry
point for users and also to be used within CI pipelines of other
projects.

The same Dockerfile configuration was used
in
https://build.opensuse.org/package/view_file/home:okurz:container/isotovideo-qemu-x86/Dockerfile?expand=1
which in turn was used in
https://github.com/os-autoinst/os-autoinst-distri-example/pull/4
to show how isotovideo can be run as part of a CI pipeline using github
actions.

Related progress issue: https:/progress.opensuse.org/issues/77905