t/22-svirt: Fix sporadic failures of `run_ssh_cmd(keep_open=>0)`
Ticked: poo#69346
The assumption is, that the NET:SSH2 module calls DESTROY of one of the
created objects and then reuse the same memory (see _new implementation
in SSH2.xs [1]).
The "fix" is, do not use `_new` to create the class anymore. Just
blessing a HASH. And create a unique random string to identify each
instance. This identifier is used for the key in `$ssh_obj_data`.

This change is actually the minimal impact solution I found. Because we do
not use the internal `_new()` method of NET:SSH2 anymore, but getting a blessed
HASH, we could improve this code later and drop `$ssh_obj_data` completly.

[1] https://metacpan.org/source/RKITOVER/Net-SSH2-0.71/SSH2.xs#L902