Handle PRODUCTDIR logic here and find needles under DEFCASEDIR
This pull request is a part of https://github.com/os-autoinst/openQA/pull/3604.
Because PRODUCTDIR is a child directory of CASEDIR, we simply find
`main.pm` throught out CASEDIR.
When CASEDIR is a git repository URL of a test distribution, there is no
needle directory under CASEDIR, so we need to use `DEFCASEDIR` to find
the correct needle directory. Because `DEFCASEDIR` is the true test
distribution path which a worker uses.