Add better tests for lockapi
The existing unit tests were not able to reveal the last regression
and need to be improved. This changes adds new unit tests for lockapi
to 30-mmapi.t to tests these related APIs in the same way and removes
the old tests (09-lockapi.t).

See https://progress.opensuse.org/issues/80264