Use trivial 'prove' call in wrapper Makefile to fully replace invoke-tests
There should be no need for an additional wrapper shell script when we
call prove how it is supposed to be called in combination with a simple
"test" make target in Makefile which can be used by "cover".

"cover" from Devel::Cover is designed to use a (fake?) Makefile. As
alternative we can also give the test command as argument to the `-make`
option of cover.

More details about that I found in
https://github.com/pjcj/Devel--Cover/issues/192
and
https://stackoverflow.com/questions/11212425/how-to-use-develcover-with-prove/28436314#28436314

These changes also fix the coverage reporting when triggering tests from
the main repository folder and including the coverage data for t/ as we
want to achieve.