tools: Add prove arg --timer to be consistent with openQA
Same as for openQA I consider it worthwhile to call prove with the
"--timer" argument so that we get a report about the runtime of
individual test modules, especially from CI runs.

This commit adds the prove arg "--timer" as default value for the
variable PROVE_ARGS which is used within tools/invoke-tests where prove
is called.