tools/tidy: Use the perltidy version from the calling repo
tools/tidy is linked into various other repos as well, which use it for running
perltidy on the contained code. Those repos have their own cpanfile which may
specify a different version of Perl::Tidy, which should get used instead.
This is especially important if os-autoinst is always pulled from master, as
changes in its cpanfile otherwise affect other repos as well.

If the repo tools/tidy is run against does not specify a version of Perl::Tidy,
it falls back to the old behaviour.