14-isovideo.t: create 'foo' branch in test git repo
The test uses this branch, but doesn't create it when setting up
the repo. Somehow this seems to work most of the time - I'm not
sure how - but fails on package builds in Fedora infrastructure
(it works doing a package build locally in mock, oddly). Creating
the branch when setting up the repo seems to avoid the problem.

Signed-off-by: Adam Williamson <awilliam@redhat.com>