No longer convert to 16 colours
8bit gray scales of a blurred image is already weak enough, the
conversation to 16 shades made it just too weak to detect certain
situations (like in poo#89314)

This comes at a price obviously as other screens that only differ
within these 16 shades of grey will now drop below the threshold
and will require a new needle.

But I think that the need for a new needle is easier to handle
by the tester than a wrong match.