Prevent Inline::Python failing in downstream projects
E.g. see
https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/12238/checks?check_run_id=2286115726#step:5:32
failing with

```
! Configure failed for Inline-Python-0.56. See /github/home/.cpanm/work/1617787806.1116/build.log for details.
```

I would not have expected that os-autoinst-distri-opensuse CI checks try
to install that package as was listed as only a "test" requirement in
cpanfile#L75 and the actual error why it fails to be pulled in is not
obvious. We should look into both errors but to avoid this for now this
commit is moving the dependency into the dependency list which is really
only used within the local test environment of os-autoinst.

Related progress issue: https://progress.opensuse.org/issues/90761