Fix return value of _dbus_do_call
This was relying on perl's default return behaviour before
(return the result of the final line of the function, if it's an
expression). But now that's wrong, as the last line is something
else. We need to be explicit about what we're returning (which
would generally be a good thing to do anyway).

Related progress issue: https://progress.opensuse.org/issues/91163

Signed-off-by: Adam Williamson <awilliam@redhat.com>