console_proxy: Proxy also the caller context (list or scalar)
The caller context get lost with current console_proxy implementation.
This force each call into SCALAR context, even if the actual console
function return a list.

With this commit, we solve this problem and forward the `wantarray`
to the actual console function call. If we need the list context, the
result is transferred as ARRAY_REF.

This will make this generic interface really transparent when calling
console functions.

I tried to avoid the usage of `wantarray` but so far, I didn't found a
transparent solution, which works for all combinations.