Support UEFI_PFLASH_VARS is a basename
Fix a regression issue caused by
https://github.com/os-autoinst/openQA/pull/3863.

When UEFI_PFLASH_VARS is a basename, we do symlink on openQA worker side
and set the value as a basename. But we still need to use the absolute path when
doing qemu-img.

Here is an incomplete job example: https://openqa.opensuse.org/tests/1735786