Hyperv backend minor updates
With hyper2019 server we have more hyperv workers therefore we should do
cleanup as soon as the assets are not needed any more.

The second part of this PR is related to checkpoints aka snapshots in hyperv.
There are 2 types of snapshots _Standard_ and _Production_. The last one
is default setup on hyperv 2016 and 2019. AS the VM related code is
split between backend/svirt and bootloader_hyperv files, it should be
clear whether reverted checkpoint failed, VM is not starting (if
_Production_ checkpoint was created) or we have a problem to reconnect
RDP session.

- related PR:
  - [Set checkpoint "Standard" for hyperv 2016 and 2019](https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/12607)
- [[jeos][sporadic] test fails or incompletes in snapper_jeos_cli with reason auto_review:"(?s)backend died: xfreerdp did not start.*svirt"](https://progress.opensuse.org/issues/80886)
- VRs:
  - [sle-15-SP2-JeOS-for-MS-HyperV-QR-x86_64-Build15.89-jeos-filesystem@svirt-hyperv](http://kepler.suse.cz/tests/5119)
  - [sle-15-SP3-JeOS-for-MS-HyperV-x86_64-Build2.36-jeos-base+phub@svirt-hyperv-uefi ](http://kepler.suse.cz/tests/5120)