Fix test execution - force users to specify test directory
Right now, prove is being run twice: once in the t directory and once in the xt directory, in that order. That works fine when no specific tests to run are defined by the user as arguments, but when the user does provide specific tests, those might fail. This pr deals with that case.

- Related issue: https://progress.opensuse.org/issues/92497