Skip empty ISO_X variables when configuring block devices
We are currently checking whether ISO is actually defined but do not perform the
same check for ISO_X, which could be an empty string. In that case os-autoinst
calls qemu on an empty file and fails in a non-obvious way.
Instead we now skip adding ISO_X as a blockdevice if the value is not truthy.