Strip down qemu-backend symptom reporting to focus on real errors
Since c670720d instead of helpful output from qemu the log first only
records the unavailability of a QMP connection but not the real qemu
error which is handled by a different component. To not distract users
we can strip down the QMP related message a bit with no source file
location (not helpful to users) and no blank line in the middle.

Related progress issue: https://progress.opensuse.org/issues/19720