qemu: remove medium from cd when ejecting
Just ejecting the cd does not prevent the SUT to access it - the tray
can be injected again (even by the OS itself). This makes the 'eject_cd'
much less useful in tests that verify if the system indeed (or the
installer itself) works independently of the boot CD/DVD.

Improve this by executing also 'blockdev-remove-medium' QMP command,
that indeed remove the medium after ejecting the CD.

Furthermore, drop support for older 'device' parameter, as it is
obsolete now, and not supported in newer commands like
'blockdev-remove-medium'. In practice, one calls just 'eject_cd()'
anyway.