Add video-stream console
This console is meant to receive video stream captured from the SUT
using an external device (like HDMI over IP extender, or HDMI usb
dongle). This console uses ffmpeg to receive the stream and convert to a
format easy for os-autoinst to handle. Currently it has hardcoded
conversion to 2 FPS, and then os-autoinst scales it (down) to 1024x768.
I have tested it with a cheap Delock HDMI extender that uses HDbitT
"protocol". This devices sends the video stream to a multicast address
239.255.42.42:5004 and can be easily received via ffmpeg.

On its own its provide only view of the system screen, but
doesn't allow to control it (mouse/keyboard), so it isn't a full
"console".
The latter part is delegated to an external tool/script that is supposed
to handle commands on its stdin. Those should are expected to be
in `send_key` format (one per line), or
`mouse_move <x> <y>`/`mouse_button <button-mask>`. Keep in mind that
os-autoinst considers the whole screen as 1024x768, which may be
important for calculating actual movements of the mouse/table/whatever.