testapi::compat_args: Fix if given value contains invalid regex
It was discovered during a call like:
```
$instance->run_ssh_command(q(rpm -q --qf '%{VERSION}\\n' ltp));
```
while `publiccloud::instance::run_ssh_command()` used `compat_args`
like:
```
testapi::compat_args({cmd => undef}, ['cmd'], @_);
```
which makes the first argument to `cmd` if not given via as hash (e.g.
`cmd => 'exit 1'`).

But this failed with
```
  Test died: Unescaped left brace in regex is illegal here in regex; marked by <-- HERE in m/^rpm -q --qf '%{ <-- HERE VERSION}\n' ltp$/ at /usr/lib/os-autoinst/testapi.pm line 2311.
  testapi::compat_args(undef, undef, "rpm -q --qf '%{VERSION}\\n' ltp") called at sle/lib/publiccloud/instance.pm line 51
```
because we use the given argument in a regex instead of do a `eq`
compare.