Fix unintended code sections in documentation
It is necessary to separate the line `=cut` with a blank line before and
afterwards to really stop the processing of the content into
documentation. This commit fixes incorrect rendering of the test API
documentation into the openQA documentation page
http://open.qa/api/testapi

Tested locally with the commands from openQA:

```
ln -s ../os-autoinst src && tools/generate-documentation-genapi && asciidoctor testapi.asciidoc
```

and visually checking the generated HTML document.