Use tidyall for faster tidying
* Adding Test::Code::TidyAll as tidy check
* Using tidyall in tools/tidy instead of own implementation

The time for a no-op check if the local cache has been created once for
me is 0.5s vs. 16s for the previous implementation of tools/tidy. If a
file is changed then only this file will be tidied and the execution
time will be around 1-2s.

tidyall is certainly not perfect but seems to be better than our custom
implementation :)

After we implement this for os-autoinst I expect a much bigger speed
improvement for os-autoinst-distri-opensuse.

Related progress issue: https://progress.opensuse.org/issues/99678