distribution: Fix sporadic here-document input failure
We encounter a stability problem with simple `script_output('w')` call.
1 out of ~50 calls failed, because the content of the script is
```
w\r
```
instead of
```
w\n
```

The actual root cause is unknown. But it seems that we encounter similar
problems as fixed with 87f895f3. The problem is, that a `wait_screen_change()`
call around the `type_string("\n")` didn't solve the problem, as the
changed screen where sometimes detected to early. During tests, we
encountered a minimal sleep of 800ms.
To not use a native `sleep` the decision was made to use
`wait_still_screen` functionality of `type_string`. This is a minimum
sleep, but guarded that the screen did not change within this
period.

The behavior was only seen on hyperv and vmware hosts. Thus we added a
`testapi::backend_get_wait_still_screen_on_here_doc_input()` function which
return the amount of seconds to wait for still screen, before start
typing the script.

The openqa variable `_WAIT_STILL_SCREEN_ON_HERE_DOC_INPUT` can be used to
overwrite this value.

Ticked: https://progress.opensuse.org/issues/60566