Select sensible default of 5 minutes for SSH command timeout
As shown by experiments with various openQA jobs we should be fine to
select a timeout of multiple minutes by default. My tests have shown
that a timeout of 1s is too low (as expected) and 30s or 3600s works
fine so a conservative selection of 5 minutes should be ok and helpful.

Related progress issue: https://progress.opensuse.org/issues/99123