git subrepo pull (merge) --force external/os-autoinst-common
subrepo:
  subdir:   "external/os-autoinst-common"
  merged:   "08fd08b0"
upstream:
  origin:   "git@github.com:os-autoinst/os-autoinst-common.git"
  branch:   "master"
  commit:   "08fd08b0"
git-subrepo:
  version:  "0.4.3"
  origin:   "???"
  commit:   "???"

I tried to update with the "--force" parameter but with no luck:

```
$ git subrepo clean --all
Removed branch 'subrepo/external/os-autoinst-common'.
$ git subrepo pull external/os-autoinst-common
git-subrepo: There is already a worktree with branch subrepo/external/os-autoinst-common.
Use the --force flag to override this check or perform a subrepo clean
to remove the worktree.
$ git subrepo clean --ALL --force
$ git subrepo pull external/os-autoinst-common
git-subrepo: There is already a worktree with branch subrepo/external/os-autoinst-common.
Use the --force flag to override this check or perform a subrepo clean
to remove the worktree.
```