Fix missing subsecond display in video captions
https://github.com/os-autoinst/os-autoinst/pull/1059 introduced video
captions overlaying wallclock time over the video. The pull request even
had a discussion about the subsecond precision in the relative steps.
But for the actual absolute timestamp the calculation was always there
to correctly calculate subsecond precision and display as milliseconds.
However "gettimeofday()" only returns the expected float number when
used as a string. So far the code called gettimeofday as the argument
for the method "format_vtt_timestamp" meaning that gettimeofday yields a
list of two elements where we discard the second element including the
subsecond part meaning that the timestamp was always showing ".000"
instead of the real subsecond value. An example from a recent real-life
openQA job
https://openqa.opensuse.org/tests/1963236
shows

```
0
00:00:00.000 --> 00:00:00.041
[2021-10-11T12:24:23.000]

1
00:00:00.041 --> 00:00:00.083
[2021-10-11T12:24:23.000]

2
00:00:00.083 --> 00:00:00.125
[2021-10-11T12:24:23.000]

3
00:00:00.125 --> 00:00:00.166
[2021-10-11T12:24:24.000]
```

so multiple occurences of the same timestamp with the misleading and
wrong ".000" part included.

The problem was found when I was trying to introduce perl signatures
which failed because the second return value of gettimeofday was
unexpectedly discarded.

This commit adjusts the tests and changes the call to gettimeofday by
forcing it to return a string value, i.e. a float number.