Fix handling resume command (of developer mode) when not paused
This fixes errors we observe in the fullstack test logs, see
https://progress.opensuse.org/issues/101734.

Note that the problem can also be reproduced manually by sending
`{"cmd": "resume_test_execution"}` via the developer mode console while the
test is *not* paused. This leads to the exact same error message than we
see in the fullstack test (and the job ends up incomplete).