Ignore the previous tests fail hook output in the search for issues
After the main test code runs, search_for_expected_serial_failures looks
through the generated serial output for certain patterns. In the case that the
test failed, the post_fail_hook runs after that, which might also produce output
on the serial console. Before this commit, this was not skipped and so the call
to search_for_expected_serial_failures on the next module sees it.

This can even have a cascading effect if there's a OOM message in the log files,
if it was missed in an earlier module. Then the post_fail_hook would print that
to the serial console and thus cause the next module to fail, which in turn
prints the message again and the cycle repeats.

To simplify the code, tie reading new serial data to setting the offset.