Delete unused exception handling in backend::qemu
3ec3779b changed the method "select_console" to return an error object
instead of throwing an exception so the previously used exception
handling in backend::qemu was ineffective. So far in case of errors
mostly isotovideo stops with an error from myjsonrpc so termination is
ensured anyway :) As the error handling code was effectively dead and
only the case of not alive qemu was handled I chose to delete the block.
As alternative we could check the return value of `select_console` and
act accordingly but from within the qemu backend we can not do much
anyway as we do not control all the other processes.