baseclass: Fix errors in some tests by increasing _CHKSEL_RATE_HITS
Ticket: https://progress.opensuse.org/issues/60161

This just double the allowed read attempts in 30s to 30k. This is needed, as
faster workers can process more then 15k read attempts on one socket in
30s. In general idea of accounting reads per socket isn't a good.
But we do, and the default was to low these days.
Of course all tests, which do heavy console interaction, can set this value
for it's own, but I think it's not obvious that a console error might be this
variable. So having a better default value makes sense.

We can easily trigger up to 55k by a simple valid testcode like this
```perl
while(1) { script_output('true', quiet => 1) };
```