Add support for v4l2 device as an input to video-stream console
This adds necessary support for TC358743 HDMI-to-CSI device connected to
RaspberryPi 4. The benefit over HDMI-over-IP converter is one less box
and less stress on the network.

ffmpeg is generally happy about '-i /dev/video0', but the device itself
needs a little extra setup:
1. At the beginning, EDID needs to be loaded. I recommend one from
   https://github.com/6by9/CSI2_device_config, but in principle any
   should work. Note on RPi4 there is a bandwidth limitation (default 2
   lanes, not all 4), which means 1920x1080@50Hz works but
   1920x1080@60Hz does not.
2. On every resolution change (including the initial one), timings needs
   to be reconfigured. This requires stopping ffmpeg that reads from the
   device. I haven't found an elegant way for getting resolution change
   event, so the code polls for it every 3s.

The need for the second point is detected automatically. EDID needs to
be provided via GENERAL_HW_EDID, but can be also loaded independently of
openQA (like with a system startup script).