Do not limit saving assets to qemu backend only
This can be implemented in other backends too, do not silently ignore
it. If saving assets is requested with relevant variable, but not
supported by the backend (`do_extract_assets` is not implemented), it
will result in an exception instead of silently ignoring it.

Silent ignore can be especially frustrating if another job depends on
such asset and is using START_AFTER_TEST. It will seemingly work fine,
but the later job will use outdated or different disk image.