baseclass: Improve error message on "half-open socket"
Ticket: https://progress.opensuse.org/issues/60161

The current error message is miss leading. As we die here because of
some read attempts threshold. And it must not be correlated to the
current active console.

This try to improve the error message and point to the actual fd which
hit the limit. Also we can name the fd's so it is more easy to identify
them on debugging.