Avoid running into TBB's soft limit for the max. number of threads
When executing the signalblocker test in Factory builds (which use
`libtbb12-2021.4.0`) we currently see the warning

```
TBB Warning: The number of workers is currently limited to 5. The request for 7 workers is ignored. Further requests for more workers will be silently ignored until the limit changes.
```

and the test deadlocks.

This change should avoid it by staying below this limit.

See https://progress.opensuse.org/issues/103029