xt: Fix style check when called in CI from other directory
From
https://github.com/os-autoinst/os-autoinst/runs/4326980124?check_suite_focus=true#step:3:624
and in every os-autoinst CI test run:

```
8: [18:02:43]
8: All tests successful.
8: Files=49, Tests=1108, 415 wallclock secs ( 0.68 usr  0.11 sys +
266.31 cusr 21.89 csys = 288.99 CPU)
8: Result: PASS
8: fatal: ambiguous argument ':!COPYING': unknown revision or path not
in the working tree.
8: Use '--' to separate paths from revisions, like this:
8: 'git <command> [<revision>...] -- [<file>...]'
8: fatal: ambiguous argument ':!COPYING': unknown revision or path not
in the working tree.
8: Use '--' to separate paths from revisions, like this:
8: 'git <command> [<revision>...] -- [<file>...]'
8: fatal: ambiguous argument ':!COPYING': unknown revision or path not
in the working tree.
8: Use '--' to separate paths from revisions, like this:
8: 'git <command> [<revision>...] -- [<file>...]'
8: [18:02:44] ./01-style.t ............. ok      406 ms ( 0.00 usr  0.00
sys +  0.33 cusr  0.07 csys =  0.40 CPU)
```

The problem is caused due to our tests in CI (github actions) to be
called from another directory /opt/build which seems to be within a git
repository itself but is not the working copy of os-autoinst.
I could reproduce the problem locally with a call

```
$ (cd ~ && ~/local/os-autoinst/os-autoinst/xt/01-style.t)
fatal: ambiguous argument ':!COPYING': unknown revision or path not in
the working tree.
```

This commit fixes xt/01-style.t by moving into the according working
directory as required. The same approach is done in other tests, e.g.
t/00-compile-test.t

Related progress issue: https://progress.opensuse.org/issues/103329